+++ src/components/app/app.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {Provider} from 'react-redux';
 import renderer from 'react-test-renderer';
 
-import App from './app.jsx';
+import App from './app';
 
 import {store} from '../../test-data/store';
