+++ src/components/map/map.e2e.test.js
@@ -3,7 +3,7 @@ import React from 'react';
 import {configure, shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 
-import {Map} from './map.jsx';
+import {Map} from './map';
 
 import {offers} from '../../test-data/index';
