+++ src/components/map/map.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {Provider} from 'react-redux';
 import renderer from 'react-test-renderer';
 
-import Map from './map.jsx';
+import Map from './map';
 
 import {offers} from '../../test-data';
 import {store} from '../../test-data/store';
