+++ src/components/no-offers/no-offers.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import NoOffers from './no-offers.jsx';
+import NoOffers from './no-offers';
 describe(`NoOffers`, () => {
   it(`NoOffers component should render correctly`, () => {
     const tree = renderer.create(
