+++ src/components/no-offers/no-offers.tsx
