+++ src/components/offer-card/offer-card.e2e.test.js
@@ -7,7 +7,7 @@ import {ActionCreator as MapActionCreator} from '../../actions/map-actions';
 import {offers} from '../../test-data';
 import {store} from '../../test-data/store';
 
-import OfferCard from './offer-card.jsx';
+import OfferCard from './offer-card';
 import {Pages} from '../../const';
 import {AuthorizationStatus} from '../../reducers/user/user-reducer';
