+++ src/components/offer-card/offer-card.tsx
