+++ src/components/offers-list/offers-list.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import OfferCard from '../offer-card/offer-card.jsx';
+import OfferCard from '../offer-card/offer-card';
 
 import {offerPropType} from '../../types';
 import {Pages} from '../../const.js';
