+++ src/components/sort-menu/sort-menu.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import SortMenu from './sort-menu.jsx';
+import SortMenu from './sort-menu';
 
 import {SortType} from '../../const.js';
