+++ src/components/spinner/spinner.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import Spinner from './spinner.jsx';
+import Spinner from './spinner';
 
 describe(`Spinner`, () => {
