+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
 
-import App from './components/app/app.jsx';
+import App from './components/app/app';
 
 import {store} from './store';
